+++ src/utils.cuh
@@ -24,8 +24,8 @@ void initialize_matrix(std::vector<T> &matrix, int rows, int cols) {
   for (int i = 0; i < rows * cols; ++i) {
     if constexpr (std::is_same_v<T, float>) {
       matrix[i] = distribution(generator);
-    } else if constexpr (std::is_same_v<T, bf16>) {
-      matrix[i] = __float2bfloat16(distribution(generator));
+    } else if constexpr (std::is_same_v<T, half>) {
+      matrix[i] = __float2half(distribution(generator));
     }
   }
 }
